Add lecture home controller indirection
